Add abstraction to Day 17
