+++ Day17 - Set and Forget/Day17.csproj
@@ -4,6 +4,11 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <Nullable>enable</Nullable>
+    <RootNamespace>AdventOfCode.Year2019.Day17</RootNamespace>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="mMosiur.AdventOfCode.Abstractions" Version="1.0.1" />
+  </ItemGroup>
 
 </Project>
