+++ Day17 - Set and Forget/IntcodeMachine/Instructions/InstructionImplementations/OutputInstruction.cs
@@ -1,6 +1,6 @@
 using System.IO;
 
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 public partial class IntcodeMachine
 {
