+++ Day17 - Set and Forget/IntcodeMachine/Instructions/InstructionParameters/InstructionParameter.cs
@@ -1,4 +1,4 @@
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 public partial class IntcodeMachine
 {
