+++ Day17 - Set and Forget/IntcodeMachine/IntcodeMachineMemory.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 public partial class IntcodeMachine
 {
