+++ Day17 - Set and Forget/IntcodeMachine/OutputInstructionExecutedEventArgs.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 public class OutputInstructionExecutedEventArgs : EventArgs
 {
