+++ Day17 - Set and Forget/NoRotationAvailableException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 [Serializable]
 public class NoRotationAvailableException : Exception
