+++ Day17 - Set and Forget/Traverser.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Day17;
+namespace AdventOfCode.Year2019.Day17;
 
 public class Traverser
 {
